+++ Projects/Domains/ChartDomain/Interface/DataSource/RemoteChartDataSource.swift
@@ -1,5 +1,3 @@
-import DataMappingModule
-import DomainModule
 import ErrorModule
 import Foundation
 import RxSwift
