+++ Projects/Domains/ChartDomain/Interface/Entity/ChartRankingEntity.swift
