+++ Projects/Domains/ChartDomain/Sources/UseCase/FetchChartRankingUseCaseImpl.swift
@@ -1,7 +1,5 @@
-import DataMappingModule
-import DomainModule
-import ErrorModule
 import RxSwift
+import ChartDomainInterface
 
 public struct FetchChartRankingUseCaseImpl: FetchChartRankingUseCase {
     private let chartRepository: any ChartRepository
