+++ Projects/Features/ChartFeature/Sources/Components/ChartContentComponent.swift
@@ -3,6 +3,7 @@ import DataMappingModule
 import DomainModule
 import Foundation
 import NeedleFoundation
+import ChartDomainInterface
 
 public protocol ChartContentDependency: Dependency {
     var fetchChartRankingUseCase: any FetchChartRankingUseCase { get }
